Obtém dados de um ano.

A tibble que retorna é a versão original obtida na API, ou seja, não foi limpa.

obter_ano(
  ano,
  ano_atual = FALSE,
  download_local = TRUE,
  diretorio_download = "data-raw/"
)

Arguments

ano
Numérico. Qual é o ano que deve ser feito o download. Ex: 2019.

ano_atual
Lógico. Se quiser obter dados de anos anteriores, deve ser FALSE. Se quiser obter dados do ano atual, deve ser TRUE. Por padrão, é FALSE.

download_local
Lógico. Se for TRUE, uma cópia dos dados baixados será salvo localmente.

diretorio_download
Diretório onde o arquivo final deve ser salvo. Ex: "data-raw/"

Value

Retorna um tibble com os dados do ano.

Details

Essa função demora alguns minutos para executar.

Examples

 mananciais:::obter_ano(2019, download_local = FALSE)

#> # A tibble: 2,555 x 20
#>    dia   sistema_id nome  volume_porcenta~ volume_porcenta~ volume_variacao~
#>    <chr> <chr>      <chr> <chr>                       <dbl> <chr>           
#>  1 2019~ 0          Cant~ 39,5                         39.5 0,0             
#>  2 2019~ 1          Alto~ 50,3                         50.3 -0,1            
#>  3 2019~ 2          Guar~ 59,9                         59.9 -0,4            
#>  4 2019~ 3          Cotia 43,1                         43.1 -0,1            
#>  5 2019~ 4          Rio ~ 82,0                         82.0 -0,3            
#>  6 2019~ 5          Rio ~ 80,4                         80.4 0,0             
#>  7 2019~ 17         São ~ 66,2                         66.2 -0,3            
#>  8 2019~ 0          Cant~ 39,5                         39.5 0,0             
#>  9 2019~ 1          Alto~ 50,1                         50.1 -0,2            
#> 10 2019~ 2          Guar~ 59,6                         59.6 -0,3            
#> # ... with 2,545 more rows, and 14 more variables: volume_variacao_num <dbl>,
#> #   volume_operacional <dbl>, image_prec_dia <chr>, prec_dia <chr>,
#> #   prec_mensal <chr>, prec_hist <chr>, link_grafico <lgl>,
#> #   link_informacoes <lgl>, indicador_volume_dia <int>, indicador_volume <int>,
#> #   circle_image <lgl>, link_tempo <lgl>, data_original$value <chr>,
#> #   data <date>

 mananciais:::obter_ano(2020, ano_atual = TRUE, download_local = FALSE)

#> # A tibble: 1,897 x 20
#>    dia   sistema_id nome  volume_porcenta~ volume_porcenta~ volume_variacao~
#>    <chr> <chr>      <chr> <chr>                       <dbl> <chr>           
#>  1 2020~ 0          Cant~ 40,3                         40.3 -0,1            
#>  2 2020~ 1          Alto~ 76,6                         76.6 -0,2            
#>  3 2020~ 2          Guar~ 64,0                         64.0 -0,5            
#>  4 2020~ 3          Cotia 73,4                         73.4 -0,4            
#>  5 2020~ 4          Rio ~ 82,6                         82.6 -0,3            
#>  6 2020~ 5          Rio ~ 97,6                         97.6 -0,5            
#>  7 2020~ 17         São ~ 58,8                         58.8 -0,1            
#>  8 2020~ 0          Cant~ 40,4                         40.4 0,1             
#>  9 2020~ 1          Alto~ 76,4                         76.4 -0,2            
#> 10 2020~ 2          Guar~ 63,6                         63.6 -0,4            
#> # ... with 1,887 more rows, and 14 more variables: volume_variacao_num <dbl>,
#> #   volume_operacional <dbl>, image_prec_dia <chr>, prec_dia <chr>,
#> #   prec_mensal <chr>, prec_hist <chr>, link_grafico <lgl>,
#> #   link_informacoes <lgl>, indicador_volume_dia <int>, indicador_volume <int>,
#> #   circle_image <lgl>, link_tempo <lgl>, data_original$value <chr>,
#> #   data <date>